+++ packages/components/src/index.ts
@@ -20,7 +20,7 @@ export type {
   TimeInputProps,
   CheckBoxProps,
   RadioButtonGroupProps,
-  SelectProps,
+  SelectInputProps,
 } from './form'
 export {
   Form,
@@ -33,7 +33,7 @@ export {
   Password,
   RadioButton,
   RadioButtonGroup,
-  Select,
+  SelectInput,
   Text,
   TextArea,
   TimeInput,
@@ -83,3 +83,15 @@ export { Modal } from './modal'
 
 export type { SortableListProps, SortableItemProps, DraggableProps, DroppableProps } from './dnd'
 export { SortableItem, SortableList, DragDropProvider, Draggable, Droppable, arrayMove } from './dnd'
+
+export type { RadioProps, RadioGroupProps } from './radio'
+export { Radio, RadioGroup } from './radio'
+
+export type { DividerProps } from './divider'
+export { Divider } from './divider'
+
+export type { SelectProps } from './select'
+export { Select } from './select'
+
+export type { CheckboxGroupProps } from './checkboxGroup'
+export { CheckboxGroup } from './checkboxGroup'