+++ packages/components/src/styles/defaults.ts
@@ -7,11 +7,15 @@ import type {  AccordionStyles,
   FormTitleStyles,
   IconStyles,
   RadioButtonGroupStyles,
-  SelectStyles,
+  SelectInputStyles,
   TableStyles, ListStyles,
   CheckboxStyles,
   ModalStyles,
-  DndStyles
+  DndStyles,
+  RadioGroupStyles,
+  SelectStyles,
+  DividerStyles,
+  CheckboxGroupStyles
 } from '.'
 
 export type StclandStyles = {
@@ -24,8 +28,12 @@ export type StclandStyles = {
   accordion?: Partial<AccordionStyles>
   list?: Partial<ListStyles>
   checkbox?: Partial<CheckboxStyles>
+  checkboxGroup?: Partial<CheckboxGroupStyles>
   modal?: Partial<ModalStyles>
   dnd?: Partial<DndStyles>
+  radioGroup?: Partial<RadioGroupStyles>
+  select?: Partial<SelectStyles>
+  divider?: Partial<DividerStyles>
 
    // Form styles
    form?: Partial<FormStyles>;
@@ -37,7 +45,7 @@ export type StclandStyles = {
    number?: Partial<FormInputStyles>;
    password?: Partial<FormInputStyles>;
    radio?: Partial<RadioButtonGroupStyles>;
-   select?: Partial<SelectStyles>;
+   selectInput?: Partial<SelectInputStyles>;
    text?: Partial<FormInputStyles>;
    textArea?: Partial<FormInputStyles>;
    time?: Partial<FormInputStyles>;
@@ -101,7 +109,7 @@ export const defaultStyles: StclandStyles = {
   accordion: {
     accordion: { root: 'border rounded bg-gray-825' },
     accordionSummary: {
-      root: 'flex justify-between items-center p-4 cursor-pointer bg-gray-825 text-gray-100 border-b border-primary-main',
+      root: 'flex justify-between items-center p-4 bg-gray-825 text-gray-100 border-b border-primary-main',
       icon: 'ml-2 transition-transform duration-300 ease-in-out',
     },
     accordionDetails: { root: 'bg-gray-825 text-gray-400' },
@@ -209,7 +217,7 @@ export const defaultStyles: StclandStyles = {
     error: 'text-red-300 italic text-sm',
     radioButton: 'accent-sky-500 h-4 w-4',
   },
-  select: {
+  selectInput: {
     root: 'flex flex-col gap-2',
     label: '',
     menuButton:
@@ -236,6 +244,7 @@ export const defaultStyles: StclandStyles = {
     fieldset: 'grid gap-2 disabled:cursor-disabled grid-cols-4',
   },
   checkbox: {
+    container: 'flex flex-row',
     root: 'cursor-pointer',
     rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-400',
     primary: 'bg-primary-main border-primary-main',
@@ -251,8 +260,13 @@ export const defaultStyles: StclandStyles = {
     mdChecked: 'text-md',
     lgChecked: 'text-lg',
     disabled: 'bg-gray-300 border-gray-300 text-gray-400 hover:bg-gray-350',
-    checked: 'text-white',
-    indeterminate: 'absolute w-3/4 h-0.5',
+    indeterminatePrimary: 'absolute w-3/4 h-0.5',
+    indeterminateSecondary: 'absolute w-3/4 h-0.5',
+    indeterminateNeutral: 'absolute w-3/4 h-0.5',
+    labelPrimary: 'text-primary-main',
+    labelSecondary: 'text-secondary-main',
+    labelNeutral: 'text-gray-700',
+    labelCustomIcon: 'text-gray-925'
   },
   modal : {
     root: 'fixed inset-0 z-50 flex items-center justify-center',
@@ -286,6 +300,74 @@ export const defaultStyles: StclandStyles = {
       dragging: 'opacity-50',
       content: 'flex-1',
       removeButton: 'w-6 h-6 text-gray-300 hover:text-gray-500'
+    }
+  },
+  radioGroup: {
+    radio: {
+      root: 'flex items-center gap-2',
+      inputRoot: 'sr-only',
+      radio: 'w-4 h-4 rounded-full border-2 flex justify-center items-center border-gray-825 bg-none',
+      radioDisabled: 'border-gray-400 bg-gray-100 cursor-not-allowed',
+      selected: 'bg-primary-main',
+      innerCircle: 'w-1.5 h-1.5 rounded-full bg-gray-825',
+      text: 'text-sm text-gray-825',
+      textDisabled: 'text-gray-400'
+    },
+    radioGroup: {
+      root: 'flex',
+      vertical: 'flex-col space-y-2',
+      horizontal: 'flex-row space-x-4'
+    }
+  },
+  select: {
+    root: 'relative w-64',
+    label: 'block mb-2 text-sm font-medium text-gray-700',
+    button: 'flex flex-row items-center justify-between w-full px-4 py-2 text-left bg-neutral-900 border border-gray-400 text-neutral-400 rounded focus:outline-none',
+    menu: 'absolute z-10 w-full bg-zinc-700 border border-gray-300 text-neutral-400 rounded shadow-lg max-h-60 overflow-auto',
+    disabled: 'disabled:bg-gray-400',
+    optionContainer: 'flex items-center p-2',
+    listItem: {
+      base: 'ml-10',
+      selected: 'ml-0',
+    },
+    selectedDefaultIcon: 'h-4.5 w-4.5 ml-2 mr-3 stroke-2'
+  },
+  divider: {
+    root: 'border-0',
+    horizontal: 'w-full border-t',
+    vertical: 'h-full border-l',
+    thin: 'border-[1px]',
+    medium: 'border-[2px]',
+    thick: 'border-[4px]',
+  },
+  checkboxGroup: {
+    root: 'flex',
+    vertical: 'flex-col gap-2',
+    horizontal: 'flex-row gap-4',
+    checkbox: {
+      container: 'flex flex-row items-center gap-2',
+      root: 'cursor-pointer',
+      rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-400',
+      primary: 'bg-zinc-900 border-zinc-900 text-white',
+      secondary: 'bg-yellow-400 border-yellow-600 text-zinc-900',
+      neutral: 'bg-gray-600 border-gray-600 text-white',
+      uncheckedPrimary: 'border-zinc-900',
+      uncheckedSecondary: 'border-yellow-600',
+      uncheckedNeutral: 'border-gray-600',
+      sm: 'w-5 h-5',
+      md: 'w-6 h-6',
+      lg: 'w-8 h-8',
+      smChecked: 'text-xs',
+      mdChecked: 'text-md',
+      lgChecked: 'text-lg',
+      disabled: 'bg-gray-300 border-gray-300 text-gray-400 hover:bg-gray-350 cursor-auto',
+      indeterminatePrimary: 'absolute w-3/4 h-0.5 bg-primary-main',
+      indeterminateSecondary: 'absolute w-3/4 h-0.5',
+      indeterminateNeutral: 'absolute w-3/4 h-0.5',
+      labelPrimary: 'text-zinc-900',
+      labelSecondary: 'text-yellow-200',
+      labelNeutral: 'text-gray-700',
+      labelCustomIcon: 'text-gray-925'
     }
   }
 }