+++ packages/components/src/withStyles/index.ts
@@ -6,6 +6,11 @@ export { List } from '../list/withStyles'
 export { Checkbox } from '../checkbox/withStyles'
 export { Modal } from '../modal/withStyles'
 export { DragDropProvider } from '../dnd/withStyles'
+export { RadioGroup } from '../radio/withStyles'
+export { Radio } from '../radio/withStyles'
+export { Select } from '../select/withStyles'
+export { Divider } from '../divider/withStyles'
+export { CheckboxGroup } from '../checkboxGroup/withStyles'
 
 export {
   Form,
@@ -16,7 +21,7 @@ export {
   NumberInput,
   Password,
   RadioButtonGroup,
-  Select,
+  SelectInput,
   Text,
   TextArea,
   TimeInput,